rename {SIMDSupport => simd_support}.hpp
