+++ include/HIRT_Multichannel_Convolution/ConvolvePartitioned.hpp
@@ -5,7 +5,8 @@
 #include "../HISSTools_FFT/HISSTools_FFT.hpp"
 
 #include "ConvolveUtilities.hpp"
-#include "../SIMDSupport.hpp"
+#include "../simd_support.hpp"
+
 
 #include <algorithm>
 #include <cstdint>
