+++ include/HISSTools_FFT/HISSTools_FFT_Core.hpp
@@ -8,7 +8,8 @@
 #include <cstdlib>
 #include <functional>
 
-#include "../SIMDSupport.hpp"
+#include "../simd_support.hpp"
+
 
 // Type definitions for Apple / HISSTools FFT
