+++ include/simd_support.hpp
